hex-mdev: fix remaining issues with fence and barrier clearing in CPY_FENCE
