+++ ggml/src/ggml-hexagon/htp/cpy-ops.c
@@ -434,18 +434,24 @@ int op_cpy(struct htp_ops_context * octx) {
     int status = exec_cpy(octx, &use_dma);
 
     if (octx->op == HTP_OP_CPY_FENCE) {
-        const struct htp_tensor * sync = octx->src[1];
-        assert(sync && (sync->flags & HTP_TENSOR_FENCE));
         if (status == HTP_STATUS_OK && !use_dma) {
             // htp_tensor_flush_all(octx->ctx, octx->dsts, 1);
             qurt_mem_cache_clean((qurt_addr_t) 0, 0, QURT_MEM_CACHE_FLUSH_INVALIDATE_ALL, QURT_MEM_DCACHE);
         }
 
-        const uint32_t seq = (uint32_t) octx->op_params[0];
-        atomic_uint * sync_fence = (atomic_uint *) (uintptr_t) sync->data;
-        htp_fence_write(sync_fence, seq, status);
+        if (octx->ctx->mdev.count > 1) {
+            status = htp_mdev_group_barrier(octx, status);
+        }
 
-        FARF(HIGH, "ggml-hex: sync-release : fence %p seq %u status %d\n", sync_fence, seq, status);
+        if (octx->ctx->mdev.idx == 0) {
+            const struct htp_tensor * sync = octx->src[1];
+            assert(sync && (sync->flags & HTP_TENSOR_FENCE));
+            const uint32_t seq = (uint32_t) octx->op_params[0];
+            atomic_uint * sync_fence = (atomic_uint *) (uintptr_t) sync->data;
+            htp_fence_write(sync_fence, seq, status);
+
+            FARF(HIGH, "ggml-hex: sync-release : fence %p seq %u status %d\n", sync_fence, seq, status);
+        }
     }
 
     return status;
