+++ ggml/src/ggml-hexagon/htp/main.c
@@ -1013,69 +1013,6 @@ static void mdev_group_init(struct htp_context * ctx, const struct htp_opbatch_r
     ctx->mdev.fence_seq = (uint32_t)((req->seq & 0xfffff) << 12);
 }
 
-static int mdev_group_fence(struct htp_ops_context * octx, int op_status) {
-    struct htp_context * ctx = octx->ctx;
-    if (ctx->mdev.count <= 1) {
-        return op_status;
-    }
-
-    assert(ctx->mdev.fence_base != NULL);
-
-    const uint32_t seq = ++ctx->mdev.fence_seq;
-
-    struct htp_thread_trace * tr = &ctx->trace[0];
-    htp_trace_event_start(tr, HTP_TRACE_EVT_FENCE, (uint16_t) seq);
-
-    const uint32_t mdev_idx   = ctx->mdev.idx;
-    const uint32_t mdev_count = ctx->mdev.count;
-
-    uint8_t * fence_base   = ctx->mdev.fence_base;
-    atomic_uint * my_fence = htp_mdev_fence(fence_base, mdev_idx);
-    htp_fence_write(my_fence, seq, op_status);
-
-    if (op_status > HTP_STATUS_OK) {
-        htp_trace_event_stop(tr, HTP_TRACE_EVT_FENCE, (uint16_t) seq);
-        return op_status;
-    }
-
-    for (uint32_t d = 0; d < mdev_count; d++) {
-        if (d == mdev_idx) continue;
-        atomic_uint * peer_fence = htp_mdev_fence(fence_base, d);
-        uint64_t spins = 0;
-        while (1) {
-            uint32_t peer_seq;
-            uint32_t peer_status;
-            htp_fence_read(peer_fence, &peer_seq, &peer_status);
-            if (peer_status > HTP_STATUS_OK) {
-                FARF(ERROR, "ggml-hex: mdev %u peer %u failed with status %u : seq 0x%08x\n",
-                     mdev_idx, d, peer_status, seq);
-                htp_fence_write(my_fence, seq, peer_status);
-                htp_trace_event_stop(tr, HTP_TRACE_EVT_FENCE, (uint16_t) seq);
-                return peer_status;
-            }
-            if ((int32_t)(peer_seq - seq) >= 0) {
-                break;
-            }
-            if (++spins == 10000) {
-                FARF(ALWAYS, "ggml-hex: mdev %u waiting for mdev %u : seq 0x%08x (b %u op %u) my-fence %p peer-fence %p peer-seq 0x%08x (diff %d)\n",
-                     mdev_idx, d, seq, seq >> 12, seq & 0xfff, my_fence, peer_fence, peer_seq, (int32_t)(peer_seq - seq));
-            }
-            if (spins > HTP_FENCE_TIMEOUT) {
-                FARF(ERROR, "ggml-hex: mdev %u timeout waiting for mdev %u : seq 0x%08x (b %u op %u) peer-fence %p peer-seq 0x%08x\n",
-                     mdev_idx, d, seq, seq >> 12, seq & 0xfff, peer_fence, peer_seq);
-                htp_fence_write(my_fence, seq, HTP_STATUS_INTERNAL_ERR);
-                htp_trace_event_stop(tr, HTP_TRACE_EVT_FENCE, (uint16_t) seq);
-                return HTP_STATUS_INTERNAL_ERR;
-            }
-            hex_pause();
-        }
-    }
-    asm volatile ("syncht" : : : "memory");
-
-    htp_trace_event_stop(tr, HTP_TRACE_EVT_FENCE, (uint16_t) seq);
-    return HTP_STATUS_OK;
-}
-
 static int proc_op_req(struct htp_ops_context * octx, struct htp_buf_desc * bufs, uint32_t n_bufs,
                        struct htp_tensor * tens, uint32_t idx, struct htp_op_desc * op) {
     memcpy(octx->op_params, op->params, sizeof(octx->op_params));
@@ -1122,14 +1059,14 @@ static int proc_op_req(struct htp_ops_context * octx, struct htp_buf_desc * bufs
             dst->ne[0], dst->ne[1], dst->ne[2], dst->ne[3]);
     }
 
-    int fence_status = mdev_group_fence(octx, HTP_STATUS_OK);
+    int fence_status = htp_mdev_group_barrier(octx, HTP_STATUS_OK);
     if (fence_status > HTP_STATUS_OK) {
         return fence_status;
     }
 
     int status = execute_op(octx);
     if (status > HTP_STATUS_OK && octx->ctx->mdev.count > 1) {
-        atomic_uint * my_fence = htp_mdev_fence(octx->ctx->mdev.fence_base, octx->ctx->mdev.idx);
+        atomic_uint * my_fence = htp_mdev_fence_slot(octx->ctx->mdev.fence_base, octx->ctx->mdev.idx);
         htp_fence_write(my_fence, octx->ctx->mdev.fence_seq, status);
     }
 
@@ -1243,7 +1180,7 @@ static void process_opbatch(struct htp_context * ctx, const struct htp_opbatch_r
     qurt_mem_cache_clean((qurt_addr_t) 0, 0, QURT_MEM_CACHE_FLUSH_INVALIDATE_ALL, QURT_MEM_DCACHE);
     htp_trace_event_stop(&ctx->trace[0], HTP_TRACE_EVT_L2FLUSH, 0);
 
-    op_status = mdev_group_fence(octx, op_status);
+    op_status = htp_mdev_group_barrier(octx, op_status);
 
     profile_stop(HTP_PROF_BASIC, &batch_prof);
